Delete MetadataType from the runtime type system
Casts to `MetadataType` will never succeed in the runtime type system. We had some.

* Making it so that `InstantiatedType` and `CanonType` derive from `DefType` instead of `MetadataType` in the runtime type system.
* Moved some things into dotfiles.
