+++ src/coreclr/nativeaot/System.Private.TypeLoader/src/System.Private.TypeLoader.csproj
@@ -32,9 +32,6 @@
     <Compile Include="$(CompilerCommonPath)\Internal\Runtime\MappingTableFlags.cs">
       <Link>MappingTableFlags.cs</Link>
     </Compile>
-    <Compile Include="$(CompilerCommonPath)\Internal\Runtime\EETypeBuilderHelpers.cs">
-      <Link>EETypeBuilderHelpers.cs</Link>
-    </Compile>
     <Compile Include="$(CompilerCommonPath)\Internal\Runtime\RuntimeConstants.cs">
       <Link>RuntimeConstants.cs</Link>
     </Compile>
@@ -84,9 +81,6 @@
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Canon\InstantiatedType.Canon.cs">
       <Link>Internal\TypeSystem\InstantiatedType.Canon.cs</Link>
     </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Canon\MetadataType.Canon.cs">
-      <Link>Internal\TypeSystem\MetadataType.Canon.cs</Link>
-    </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Canon\MethodDesc.Canon.cs">
       <Link>Internal\TypeSystem\MethodDesc.Canon.cs</Link>
     </Compile>
@@ -114,9 +108,6 @@
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\AlignmentHelper.cs">
       <Link>Internal\TypeSystem\AlignmentHelper.cs</Link>
     </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\ArrayOfTRuntimeInterfacesAlgorithm.cs">
-      <Link>Internal\TypeSystem\ArrayOfTRuntimeInterfacesAlgorithm.cs</Link>
-    </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\ArrayType.cs">
       <Link>Internal\TypeSystem\ArrayType.cs</Link>
     </Compile>
@@ -129,26 +120,14 @@
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\CastingHelper.cs">
       <Link>Internal\TypeSystem\CastingHelper.cs</Link>
     </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\ConstructedTypeRewritingHelpers.cs">
-      <Link>Internal\TypeSystem\ConstructedTypeRewritingHelpers.cs</Link>
-    </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\DefType.cs">
       <Link>Internal\TypeSystem\DefType.cs</Link>
-    </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\DefType.FieldLayout.cs">
-      <Link>Internal\TypeSystem\DefType.FieldLayout.cs</Link>
     </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\FieldDesc.cs">
       <Link>Internal\TypeSystem\FieldDesc.cs</Link>
     </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\FieldDesc.FieldLayout.cs">
-      <Link>Internal\TypeSystem\FieldDesc.FieldLayout.cs</Link>
-    </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\FieldForInstantiatedType.cs">
       <Link>Internal\TypeSystem\FieldForInstantiatedType.cs</Link>
-    </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\FieldLayoutAlgorithm.cs">
-      <Link>Internal\TypeSystem\FieldLayoutAlgorithm.cs</Link>
     </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\FunctionPointerType.cs">
       <Link>Internal\TypeSystem\FunctionPointerType.cs</Link>
@@ -159,38 +138,17 @@
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\IAssemblyDesc.cs">
       <Link>Internal\TypeSystem\IAssemblyDesc.cs</Link>
     </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\IModuleResolver.cs">
-      <Link>Internal\TypeSystem\IModuleResolver.cs</Link>
-    </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\InstantiatedMethod.cs">
       <Link>Internal\TypeSystem\InstantiatedMethod.cs</Link>
     </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\InstantiatedType.cs">
       <Link>Internal\TypeSystem\InstantiatedType.cs</Link>
     </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\InstantiatedType.Interfaces.cs">
-      <Link>Internal\TypeSystem\InstantiatedType.Interfaces.cs</Link>
-    </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\InstantiatedType.MethodImpls.cs">
-      <Link>Internal\TypeSystem\InstantiatedType.MethodImpls.cs</Link>
-    </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\Instantiation.cs">
       <Link>Internal\TypeSystem\Instantiation.cs</Link>
     </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\LayoutInt.cs">
       <Link>Internal\TypeSystem\LayoutInt.cs</Link>
-    </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\ExplicitLayoutValidator.cs">
-      <Link>Internal\TypeSystem\Common\ExplicitLayoutValidator.cs</Link>
-    </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\MetadataType.cs">
-      <Link>Internal\TypeSystem\MetadataType.cs</Link>
-    </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\MetadataType.Interfaces.cs">
-      <Link>Internal\TypeSystem\MetadataType.Interfaces.cs</Link>
-    </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\MetadataType.MethodImpls.cs">
-      <Link>Internal\TypeSystem\MetadataType.MethodImpls.cs</Link>
     </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\MethodDesc.cs">
       <Link>Internal\TypeSystem\MethodDesc.cs</Link>
@@ -203,9 +161,6 @@
     </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\NotFoundBehavior.cs">
       <Link>TypeSystem\Common\NotFoundBehavior.cs</Link>
-    </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\ResolutionFailure.cs">
-      <Link>TypeSystem\Common\ResolutionFailure.cs</Link>
     </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\ParameterizedType.cs">
       <Link>Internal\TypeSystem\ParameterizedType.cs</Link>
@@ -245,9 +200,6 @@
     </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\ThrowHelper.Common.cs">
       <Link>Internal\TypeSystem\ThrowHelper.Common.cs</Link>
-    </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\TypeSystemHelpers.cs">
-      <Link>Internal\TypeSystem\TypeSystemHelpers.cs</Link>
     </Compile>
     <Compile Include="$(CompilerCommonPath)\TypeSystem\Common\Utilities\LockFreeReaderHashtableOfPointers.cs">
       <Link>LockFreeReaderHashtableOfPointers.cs</Link>
@@ -321,18 +273,6 @@
     <Compile Include="Internal\TypeSystem\TypeSystemContext.Runtime.cs" />
     <Compile Include="$(AotCommonPath)\System\Runtime\CompilerServices\__BlockAllReflectionAttribute.cs">
       <Link>System\Runtime\CompilerServices\__BlockAllReflectionAttribute.cs</Link>
-    </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Interop\MetadataType.Interop.cs">
-      <Link>Internal\TypeSystem\Interop\MetadataType.Interop.cs</Link>
-    </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Interop\InstantiatedType.Interop.cs">
-      <Link>Internal\TypeSystem\Interop\InstantiatedType.Interop.cs</Link>
-    </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Canon\CanonTypes.Interop.cs">
-      <Link>Internal\TypeSystem\Canon\CanonTypes.Interop.cs</Link>
-    </Compile>
-    <Compile Include="$(CompilerCommonPath)\TypeSystem\Interop\MarshalAsDescriptor.cs">
-      <Link>Internal\TypeSystem\Interop\MarshalAsDescriptor.cs</Link>
     </Compile>
   </ItemGroup>
   <ItemGroup Condition="'$(Configuration)'=='Debug'">
