+++ src/coreclr/tools/Common/TypeSystem/Canon/TypeSystemContext.Canon.cs
@@ -94,7 +94,7 @@ namespace Internal.TypeSystem
         public abstract bool SupportsCanon { get; }
         public abstract bool SupportsUniversalCanon { get; }
 
-        public MetadataType GetCanonType(string name)
+        public DefType GetCanonType(string name)
         {
             switch (name)
             {
