+++ src/coreclr/tools/Common/TypeSystem/Common/InstantiatedType.cs
@@ -4,6 +4,10 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 
+#if TYPE_LOADER_IMPLEMENTATION
+using MetadataType = Internal.TypeSystem.DefType;
+#endif
+
 namespace Internal.TypeSystem
 {
     public sealed partial class InstantiatedType : MetadataType
@@ -51,22 +55,12 @@ namespace Internal.TypeSystem
 
         private MetadataType InitializeBaseType()
         {
-            var uninst = _typeDef.MetadataBaseType;
+            var uninst = _typeDef.BaseType;
 
             return (_baseType = (uninst != null) ? (MetadataType)uninst.InstantiateSignature(_instantiation, default(Instantiation)) : null);
         }
 
         public override DefType BaseType
-        {
-            get
-            {
-                if (_baseType == this)
-                    return InitializeBaseType();
-                return _baseType;
-            }
-        }
-
-        public override MetadataType MetadataBaseType
         {
             get
             {
@@ -181,7 +175,7 @@ namespace Internal.TypeSystem
             if (typicalFinalizer == null)
                 return null;
 
-            MetadataType typeInHierarchy = this;
+            DefType typeInHierarchy = this;
 
             // Note, we go back to the type definition/typical method definition in this code.
             // If the finalizer is implemented on a base type that is also a generic, then the
@@ -190,7 +184,7 @@ namespace Internal.TypeSystem
 
             while (typicalFinalizer.OwningType.GetTypeDefinition() != typeInHierarchy.GetTypeDefinition())
             {
-                typeInHierarchy = typeInHierarchy.MetadataBaseType;
+                typeInHierarchy = typeInHierarchy.BaseType;
             }
 
             if (typeInHierarchy == typicalFinalizer.OwningType)
@@ -280,74 +274,6 @@ namespace Internal.TypeSystem
             return _typeDef;
         }
 
-        // Properties that are passed through from the type definition
-        public override ClassLayoutMetadata GetClassLayout()
-        {
-            return _typeDef.GetClassLayout();
-        }
-
-        public override bool IsExplicitLayout
-        {
-            get
-            {
-                return _typeDef.IsExplicitLayout;
-            }
-        }
-
-        public override bool IsSequentialLayout
-        {
-            get
-            {
-                return _typeDef.IsSequentialLayout;
-            }
-        }
-
-        public override bool IsBeforeFieldInit
-        {
-            get
-            {
-                return _typeDef.IsBeforeFieldInit;
-            }
-        }
-
-        public override bool IsModuleType
-        {
-            get
-            {
-                // The global module type cannot be generic.
-                return false;
-            }
-        }
-
-        public override bool IsSealed
-        {
-            get
-            {
-                return _typeDef.IsSealed;
-            }
-        }
-
-        public override bool IsAbstract
-        {
-            get
-            {
-                return _typeDef.IsAbstract;
-            }
-        }
-
-        public override ModuleDesc Module
-        {
-            get
-            {
-                return _typeDef.Module;
-            }
-        }
-
-        public override bool HasCustomAttribute(string attributeNamespace, string attributeName)
-        {
-            return _typeDef.HasCustomAttribute(attributeNamespace, attributeName);
-        }
-
         public override DefType ContainingType
         {
             get
@@ -355,18 +281,6 @@ namespace Internal.TypeSystem
                 // Return the result from the typical type definition.
                 return _typeDef.ContainingType;
             }
-        }
-
-        public override MetadataType GetNestedType(string name)
-        {
-            // Return the result from the typical type definition.
-            return _typeDef.GetNestedType(name);
-        }
-
-        public override IEnumerable<MetadataType> GetNestedTypes()
-        {
-            // Return the result from the typical type definition.
-            return _typeDef.GetNestedTypes();
         }
 
         public override TypeDesc UnderlyingType
