+++ src/coreclr/tools/Common/TypeSystem/Common/TypeSystemContext.cs
@@ -4,13 +4,16 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Reflection;
 
 using Internal.NativeFormat;
 
+#if TYPE_LOADER_IMPLEMENTATION
+using MetadataType = Internal.TypeSystem.DefType;
+#endif
+
 namespace Internal.TypeSystem
 {
-    public abstract partial class TypeSystemContext : IModuleResolver
+    public abstract partial class TypeSystemContext
     {
         public TypeSystemContext() : this(new TargetDetails(TargetArchitecture.Unknown, TargetOS.Unknown, TargetAbi.Unknown))
         {
@@ -44,38 +47,7 @@ namespace Internal.TypeSystem
             get;
         }
 
-        public ModuleDesc SystemModule
-        {
-            get;
-            private set;
-        }
-
-        protected void InitializeSystemModule(ModuleDesc systemModule)
-        {
-            Debug.Assert(SystemModule == null);
-            SystemModule = systemModule;
-        }
-
         public abstract DefType GetWellKnownType(WellKnownType wellKnownType, bool throwIfNotFound = true);
-
-        public virtual ModuleDesc ResolveAssembly(AssemblyName name, bool throwIfNotFound = true)
-        {
-            if (throwIfNotFound)
-                throw new NotSupportedException();
-            return null;
-        }
-
-        internal virtual ModuleDesc ResolveModule(IAssemblyDesc referencingModule, string fileName, bool throwIfNotFound = true)
-        {
-            if (throwIfNotFound)
-                throw new NotSupportedException();
-            return null;
-        }
-
-        ModuleDesc IModuleResolver.ResolveModule(IAssemblyDesc referencingModule, string fileName, bool throwIfNotFound)
-        {
-            return ResolveModule(referencingModule, fileName, throwIfNotFound);
-        }
 
         //
         // Array types
@@ -680,16 +652,6 @@ namespace Internal.TypeSystem
         protected internal virtual IEnumerable<MethodDesc> GetAllVirtualMethods(TypeDesc type)
         {
             return type.GetVirtualMethods();
-        }
-
-        /// <summary>
-        /// Abstraction to allow the type system context to affect the field layout
-        /// algorithm used by types to lay themselves out.
-        /// </summary>
-        public virtual FieldLayoutAlgorithm GetLayoutAlgorithmForType(DefType type)
-        {
-            // Type system contexts that support computing field layout need to override this.
-            throw new NotSupportedException();
         }
 
         /// <summary>
